git-svn-id: http://192.168.20.11/sequoiadb/trunk@10132 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ tools/deptool1/sequoiadbuninstall.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-
 processes=`ps -ef | grep '\<sdbcm\>'|grep -v '\<grep\>'|awk '{print $2}'`
 for process in ${processes}
 do
@@ -31,12 +30,53 @@ do
    kill -9 ${process}
 done
 
-processes=`ps -ef | grep '\<sequoiadb\>'|grep -v '\<grep\>'|awk '{print $2}'`
+processes=`ps -ef | grep '\<sequoiadb\>(\w\{1,\})'|grep -v '\<grep\>'|awk '{print $2}'`
 for process in ${processes}
 do
    kill -9 ${process}
 done
 
-rm -r ${1}
+if [ -n "${1}" ]; then
+   if [ -d "${1}" ]; then
+      echo "delete `hostname` ${1}"
+      rm -r ${1}
+   fi
+fi
+
+if [ -n "${2}" ]; then
+   if [ -d "${2}" ]; then
+      echo "delete `hostname` ${2}"
+      rm -r ${2}
+   fi
+fi
+
+if [ -n "${3}" ]; then
+   if [ -d "${3}" ]; then
+      echo "delete `hostname` ${3}"
+      rm -r ${3}
+   fi
+fi
+
+if [ -n "${4}" ]; then
+   if [ -d "${4}" ]; then
+      echo "delete `hostname` ${4}"
+      rm -r ${4}
+   fi
+fi
+
+if [ -n "${5}" ]; then
+   if [ -d "${5}" ]; then
+      echo "delete `hostname` ${5}"
+      rm -r ${5}
+   fi
+fi
+
+if [ -n "${6}" ]; then
+   if [ -d "${6}" ]; then
+      echo "delete `hostname` ${6}"
+      rm -r ${6}
+   fi
+fi
+
 rc=$?
 exit 0
